+++ src/main/java/agents/markets/meritOrder/books/OrderBook.java
@@ -11,13 +11,13 @@ import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
-import agents.markets.EnergyExchange;
+import agents.markets.DayAheadMarketSingleZone;
 import agents.markets.meritOrder.Bid;
 import de.dlr.gitlab.fame.communication.transfer.ComponentCollector;
 import de.dlr.gitlab.fame.communication.transfer.ComponentProvider;
 import de.dlr.gitlab.fame.communication.transfer.Portable;
 
-/** Handles a list of bids or asks at an energy {@link EnergyExchange} for a single time frame of trading
+/** Handles a list of bids or asks at an energy {@link DayAheadMarketSingleZone} for a single time frame of trading
  * 
  * @author Martin Klein, Christoph Schimeczek */
 public abstract class OrderBook implements Portable {